+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.12] - 2020-10-28
+
+- Add new endpoint and model: PriceTier
+- Endpoint GET PriceTiers
+
 ## [0.1.11] - 2020-09-21
 
 - Update limit of Comments in Customer model.